fixed missing asset metadata case
The client.api.getAssetMetadata(assetHash) call failed with
ClassCastException in the guts of the coroutine API because it did not
seem to handle the error case via the Promise.await() call. I had to
rewrite it using the Promise API which is arguably more messy but for
now it works.

Also, uninitialized funds are not shown in the fund lists since you
cannot do anything with them anyways and it just adds clutter.